+++ src/posts/post-thumbnail/post-thumbnail.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-post-thumbnail',
+  templateUrl: './post-thumbnail.component.html',
+  styleUrls: ['./post-thumbnail.component.css']
+})
+export class PostThumbnailComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}