+++ src/structure/post_grid/post_grid.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-post_grid',
+  templateUrl: './post_grid.component.html',
+  styleUrls: ['./post_grid.component.css']
+})
+export class Post_gridComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}