+++ src/structure/post_list/post_list.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-post_list',
+  templateUrl: './post_list.component.html',
+  styleUrls: ['./post_list.component.css']
+})
+export class Post_listComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}