+++ Changelog.md
@@ -54,3 +54,14 @@ Login Component
 Registration Component
 - Updated UI for registration page
   - updated registration html and css files
+
+
+
+### Changes by Omkar (21/9/2022)
+Added and modified modules 
+- Posts Module (New)
+  - added component for post card displayed on the post_list (empty static code)
+  - added component for post thumbnail for post_grid (empty static code)
+- Structure Module (Modifications)
+  - added component for post_grid to display on profile pages (empty static code)
+  - added component for post_list to display on homepage (empty static code)